Debugging updateMember

+++ Portal/php/deleteuser.php
@@ -2,19 +2,23 @@
     include_once "database.php";
     $database = new Database();
 
-    $table = $_GET['table'];
-    $id = $_GET['id'];
-
-    echo "$table<br>$id";
-
     $redirect = "../page/index.php";
 
-    if ($table == "admin") {
-        $database->removeAdmin($id);
-        $redirect = "../page/listadmin.php";
-    } else if ($table == "member") {
-        $database->removeMember($id);
-        $redirect = "../page/listmember.php";
+    if (isset($_GET['table']) && isset($_GET['id'])) {
+        $table = $_GET['table'];
+        $id = $_GET['id'];
+
+        if ($table == "admin") {
+            $database->removeAdmin($id);
+            $redirect = "../page/listadmin.php";
+        } else if ($table == "member") {
+            $database->removeMember($id);
+            $redirect = "../page/listmember.php";
+        }
+    } else {
+        $redirect = "../page/index.php";
     }
+
+
     header("Location: $redirect");
 ?>
